Adds the ability to transfer epubs
The script now checks the file extension and, if it's an epub, the
script will correctly classify the file as a book rather than a
notebook.

The script does not currently try to find the cover image in the epub
file (which is just a zip). The remarkable will generate one when the
book is first opened though.

This might be something I add in the future. This python tool extracts
cover images and there doesn't appear to be anything that couldn't be
achieved in bash.

https://github.com/marianosimone/epub-thumbnailer/blob/master/src/epub-thumbnailer.py